+++ common/src/main/java/org/geogebra/common/euclidian/draw/DrawEmbed.java
@@ -66,12 +66,12 @@ public class DrawEmbed extends Drawable implements DrawWidget, RemoveNeeded {
 	}
 
 	@Override
-	public int getWidthThreshold() {
+	public double getWidthThreshold() {
 		return EMBED_SIZE_THRESHOLD;
 	}
 
 	@Override
-	public int getHeightThreshold() {
+	public double getHeightThreshold() {
 		return EMBED_SIZE_THRESHOLD;
 	}
